go the second tab to work for R1
New input selections seem to work and plot is being generated!
some tweaking of some aesthetic things.
A few bugs seem to occur if filtering in different orders - need to make sure the options are set to correct data - maybe easiest is to ensure if anything changes that the values default to zero